fix c type translation for bytes on osx/intel

+++ lib/free-image/types/ffi.rb
@@ -1,7 +1,7 @@
 module FreeImage
   if !self.msvc?
     typedef :int32,  :bool
-    typedef :uint32, :byte
+    typedef :uint8,  :byte
     typedef :uint16, :word
     typedef :uint32, :dword
     typedef :int32,  :long
